+++ src/adapters/koinly/universal/YOUNG_PLATFORM/deposits.ts
@@ -1,0 +1,15 @@
+import { YoungPlatformMovement } from "@prisma/client";
+
+export const handle = (data: YoungPlatformMovement[]) => {
+  return data.map((o) => ({
+    Date: o.date.toISOString(),
+    "Sent Amount": o.debit,
+    "Sent Currency": o.currency,
+    "Received Amount": o.credit,
+    "Received Currency": o.currency,
+    "Fee Amount": 0,
+    "Fee Currency": o.currency,
+    Label: "Deposit",
+    Description: "Deposit",
+  }));
+};
