+++ src/selectors/YOUNG_PLATFORM/deposits.ts
@@ -1,0 +1,9 @@
+import { PrismaClient, YoungPlatformMovement } from "@prisma/client";
+
+export const getAll = (
+  prisma: PrismaClient
+): Promise<YoungPlatformMovement[]> => {
+  return prisma.youngPlatformMovement.findMany({
+    where: { txType: "DEPOSIT" },
+  });
+};
