+++ src/selectors/YOUNG_PLATFORM/index.ts
@@ -1,0 +1,6 @@
+import * as depositsSelectors from "./deposits";
+import * as withdrawalsSelectors from "./withdrawals";
+export const selectors = {
+  deposits: depositsSelectors,
+  withdrawals: withdrawalsSelectors,
+};
