+++ src/selectors/index.ts
@@ -1,0 +1,5 @@
+import { selectors as youngPlatformSelectors } from "./YOUNG_PLATFORM";
+
+export const selectors = {
+  youngPlatform: youngPlatformSelectors,
+};
